# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI English Mentor API
  version: 1.0.0
  description: Public REST and SSE bootstrap contract for web and Android V1.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /me:
    get:
      operationId: getMe
      responses:
        '200': { description: Current account, content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      operationId: updateMe
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateMeRequest' } } } }
      responses:
        '200': { description: Updated account, content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /me/consents/{type}:
    put:
      tags: [Privacy]
      operationId: setConsent
      parameters: [{ name: type, in: path, required: true, schema: { $ref: '#/components/schemas/ConsentType' } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SetConsentRequest' } } } }
      responses:
        '200': { description: Consent recorded, content: { application/json: { schema: { $ref: '#/components/schemas/Consent' } } } }
  /me/privacy-requests:
    post:
      tags: [Privacy]
      operationId: createPrivacyRequest
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreatePrivacyRequest' } } } }
      responses:
        '202': { description: Request accepted, content: { application/json: { schema: { $ref: '#/components/schemas/AsyncOperation' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /me/privacy-requests/{requestId}:
    get:
      tags: [Privacy]
      operationId: getPrivacyRequest
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      responses:
        '200': { description: Privacy request, content: { application/json: { schema: { $ref: '#/components/schemas/PrivacyRequest' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /learner-profile:
    get:
      tags: [Learner]
      operationId: getLearnerProfile
      responses:
        '200': { description: Learner profile, content: { application/json: { schema: { $ref: '#/components/schemas/LearnerProfile' } } } }
    put:
      tags: [Learner]
      operationId: putLearnerProfile
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLearnerProfile' } } } }
      responses:
        '200': { description: Updated learner profile, content: { application/json: { schema: { $ref: '#/components/schemas/LearnerProfile' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /learner-profile/memories/{memoryId}:
    delete:
      tags: [Learner]
      operationId: deleteLearnerMemory
      parameters: [{ name: memoryId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
      responses:
        '204': { description: Memory deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /review/queue:
    get:
      tags: [Learning]
      operationId: getReviewQueue
      parameters: [{ name: budgetMinutes, in: query, schema: { type: integer, minimum: 1, maximum: 60 } }]
      responses:
        '200': { description: Due review items, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ReviewItem' } } } } }
  /review/attempts:
    post:
      tags: [Learning]
      operationId: submitReviewAttempt
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewAttempt' } } } }
      responses:
        '201': { description: Attempt recorded, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewResult' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    googleOidc: { type: openIdConnect, openIdConnectUrl: 'https://accounts.google.com/.well-known/openid-configuration' }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 160 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string } }
...
    RequestId: { name: requestId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
  responses:
    Unauthorized: { description: Authentication required, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    NotFound: { description: Resource not found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Conflict: { description: State or idempotency conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    ValidationError: { description: Semantic validation failed, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    CefrLevel: { type: string, enum: [A1, A2, B1, B2, C1, C2] }
    ConsentType: { type: string, enum: [TERMS, PRIVACY, PERSONALIZATION, ANALYTICS, AI_PROCESSING, AUDIO_RETENTION, MODEL_IMPROVEMENT, MARKETING] }
...
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        correlationId: { $ref: '#/components/schemas/Uuid' }
        errors: { type: array, items: { type: object, required: [path, code], properties: { path: { type: string }, code: { type: string } } } }
...
    Me:
      type: object
      required: [id, status, locale, timeZone, roles, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        email: { type: [string, 'null'], format: email }
        status: { enum: [PENDING, ACTIVE, SUSPENDED, DELETION_PENDING, DELETED] }
        locale: { type: string }
        timeZone: { type: string }
        roles: { type: array, items: { type: string } }
        onboardingCompleted: { type: boolean }
        version: { type: integer }
    UpdateMeRequest:
      type: object
      additionalProperties: false
      properties: { locale: { type: string }, timeZone: { type: string } }
    Consent:
      type: object
      required: [type, policyVersion, granted, recordedAt]
      properties: { type: { $ref: '#/components/schemas/ConsentType' }, policyVersion: { type: string }, granted: { type: boolean }, recordedAt: { type: string, format: date-time } }
    SetConsentRequest:
      type: object
      additionalProperties: false
      required: [granted, policyVersion]
      properties: { granted: { type: boolean }, policyVersion: { type: string, maxLength: 40 } }
    CreatePrivacyRequest:
      type: object
      additionalProperties: false
      required: [type]
      properties: { type: { enum: [EXPORT, DELETE_ACCOUNT, DELETE_CONTENT, RECTIFY, RESTRICT_PROCESSING] } }
    PrivacyRequest:
      type: object
      required: [id, type, status, requestedAt]
      properties: { id: { $ref: '#/components/schemas/Uuid' }, type: { type: string }, status: { type: string }, requestedAt: { type: string, format: date-time }, completedAt: { type: [string, 'null'], format: date-time } }
    LearnerProfile:
      type: object
      required: [nativeLanguage, explanationLanguage, weeklyMinutes, mentorStyle, correctionIntensity, version]
      properties:
        nativeLanguage: { type: string }
        explanationLanguage: { type: string }
        selfAssessedLevel: { oneOf: [{ $ref: '#/components/schemas/CefrLevel' }, { type: 'null' }] }
        currentLevel: { oneOf: [{ $ref: '#/components/schemas/CefrLevel' }, { type: 'null' }] }
        weeklyMinutes: { type: integer, minimum: 10, maximum: 840 }
        mentorStyle: { enum: [supportive, neutral, demanding] }
        correctionIntensity: { type: integer, minimum: 1, maximum: 3 }
        interests: { type: array, maxItems: 20, items: { type: string, maxLength: 80 } }
        occupation: { type: [string, 'null'], maxLength: 160 }
        version: { type: integer }
    UpdateLearnerProfile: { $ref: '#/components/schemas/LearnerProfile' }
...
    ReviewItem:
      type: object
      required: [id, type, prompt, dueAt]
      properties: { id: { $ref: '#/components/schemas/Uuid' }, type: { type: string }, prompt: { type: string }, dueAt: { type: string, format: date-time } }
    ReviewAttempt:
      type: object
      additionalProperties: false
      required: [learningItemId, rating]
      properties: { learningItemId: { $ref: '#/components/schemas/Uuid' }, rating: { enum: [AGAIN, HARD, GOOD, EASY] }, response: { type: [string, 'null'], maxLength: 8000 }, responseTimeMs: { type: integer, minimum: 0 } }
    ReviewResult:
      type: object
      required: [attemptId, nextDueAt, state]
      properties: { attemptId: { $ref: '#/components/schemas/Uuid' }, nextDueAt: { type: string, format: date-time }, state: { type: string } }
...
    AsyncOperation:
      type: object
      required: [operationId, status]
      properties: { operationId: { $ref: '#/components/schemas/Uuid' }, status: { enum: [PENDING, PROCESSING, COMPLETED, FAILED] }, statusUrl: { type: string } }
...